fix: resolve rebase fallout and pre-existing test compile breaks
Rebased fix/533-536-537-539-fault-tolerance-and-auth onto upstream/main
(which had advanced with an unrelated basis-points-constant refactor
touching liquidation_handler, order_cleanup, order_handler, and
decrease_position_utils) to clear the PR's merge conflict.

- liquidation_handler: resolve the one real conflict by keeping the
  #533 delegation-to-order_handler fix and adopting upstream's new
  gmx_math::BPS_DIVISOR constant in place of the pre-fix stub's
  hardcoded 10_000, fixing the resulting u32/u128 type mismatch.
- order_cleanup: fix two pre-existing (unrelated to #536/#537) unit
  test compile breaks in the same file uncovered while re-verifying
  after the rebase: `Address::generate` used outside the scope of its
  local `use soroban_sdk::testutils::Address as _` import, and a
  contract `Error` compared directly against a `try_*` client result
  that decodes to `soroban_sdk::Error`.

All directly affected suites pass: liquidation-handler (16/16),
order-cleanup unit (3/3) and integration (7/7), order-handler (64/67 —
the 3 failures are the pre-existing heartbeat-timing breaks, verified
still present on clean upstream/main), adl-handler, fee-batch-sweeper,
and gmx-decrease-position-utils. `cargo check --workspace` is clean.

+++ contracts/liquidation_handler/src/lib.rs
@@ -420,7 +420,7 @@ impl LiquidationHandler {
             panic_with_error!(&env, Error::NotLiquidatable);
         }
 
-        let factor = (liquidation_factor_bps.min(gmx_math::BPS_DIVISOR as u32)) as i128;
+        let factor = liquidation_factor_bps.min(gmx_math::BPS_DIVISOR);
 
         // Issue #533: delegate the actual close to order_handler (positions live
         // there) instead of only computing what a close *would* look like. Every
@@ -432,7 +432,7 @@ impl LiquidationHandler {
             &market,
             &collateral_token,
             &is_long,
-            &(factor as u128),
+            &factor,
         );
 
         // Issue #614: record LIQUIDATION_KEEPER activity so check_keeper_heartbeat
@@ -450,7 +450,7 @@ impl LiquidationHandler {
             market,
             collateral_token,
             is_long,
-            liquidation_factor_bps: factor as u128,
+            liquidation_factor_bps: factor,
             liquidated_size_usd: result.liquidated_size_usd as u128,
             remaining_size_usd: result.remaining_size_usd as u128,
             liquidation_fee: result.keeper_execution_fee as u128,
